refactor(channels): 将 Gemini 模型选择相关组件从 conversation 移至 channels
