+++ src/renderer/pages/settings/channels/components/TelegramConfigForm.tsx
@@ -12,8 +12,8 @@ import { useTranslation } from 'react-i18next';
 import type { IChannelPairingRequest, IChannelPluginStatus, IChannelUser } from '@/channels/types';
 import { acpConversation, channel } from '@/common/ipcBridge';
 import { ConfigStorage } from '@/common/storage';
-import GeminiModelSelector from '@/renderer/pages/conversation/gemini/GeminiModelSelector';
-import type { GeminiModelSelection } from '@/renderer/pages/conversation/gemini/useGeminiModelSelection';
+import GeminiModelSelector from '@/renderer/pages/settings/channels/components/GeminiModelSelector';
+import type { GeminiModelSelection } from '@/renderer/pages/settings/channels/hooks/useGeminiModelSelection';
 import { CHANNEL_DEFAULT_AGENT_BACKEND, type AcpBackendAll } from '@/types/acpTypes';
 import { useAppMode } from '@/renderer/hooks/useAppMode';
