+++ src/renderer/pages/settings/channels/components/WeChatConfigForm.tsx
@@ -13,8 +13,8 @@ import type { IChannelPluginStatus } from '@/channels/types';
 import { openExternalUrl } from '@/renderer/utils/platform';
 import { acpConversation, channel } from '@/common/ipcBridge';
 import { ConfigStorage } from '@/common/storage';
-import GeminiModelSelector from '@/renderer/pages/conversation/gemini/GeminiModelSelector';
-import type { GeminiModelSelection } from '@/renderer/pages/conversation/gemini/useGeminiModelSelection';
+import GeminiModelSelector from '@/renderer/pages/settings/channels/components/GeminiModelSelector';
+import type { GeminiModelSelection } from '@/renderer/pages/settings/channels/hooks/useGeminiModelSelection';
 import { CHANNEL_DEFAULT_AGENT_BACKEND, type AcpBackendAll } from '@/types/acpTypes';
 import { useAppMode } from '@/renderer/hooks/useAppMode';
