+++ src/renderer/pages/settings/channels/hooks/useGeminiModelSelection.ts
